+++ buildsystem/spack/newell/modules/exago.sh
@@ -1,3 +1,3 @@
-module use -a /qfs/projects/exasgd/src/ci-newll/ci-modules/linux-centos8-power9le
-# exago@develop%gcc@8.5.0+cuda+hiop~ipo+ipopt+mpi+python+raja~rocm build_system=cmake build_type=RelWithDebInfo cuda_arch=70 dev_path=/people/svcexasgd/gitlab/19610/spack_newell arch=linux-centos8-power9le
-module load exago-develop-gcc-8.5.0-nw34u5i
+module use -a /qfs/projects/exasgd/src/ci-newll/spack-install/ci-modules/linux-centos8-power9le
+# exago@=develop%gcc@=8.5.0+cuda+hiop~ipo+ipopt+logging+mpi+python+raja~rocm build_system=cmake build_type=MinSizeRel cuda_arch=70 dev_path=/people/svcexasgd/gitlab/22332/spack_newell generator=make arch=linux-centos8-power9le
+module load exago/develop-gcc-8.5.0-bybdibf